Added Slider to main Page
Slider component has been added into the main page when a user is logged in. New slider folder created. Images for slider can be found in /main/slider/images.
Made a slight change to the brightness of the main background - but can be reverted by changing image link in app.css.